+++ winter/src/main/kotlin/io/jentz/winter/Component.kt
@@ -26,7 +26,7 @@ class Component internal constructor(
      */
     val qualifier: Any?,
 
-    private val dependencies: Map<TypeKey, UnboundService<*, *>>,
+    private val dependencies: Map<TypeKey<*, *>, UnboundService<*, *>>,
 
     /**
      * Set to true if any of the services requires lifecycle callbacks.
@@ -89,31 +89,18 @@ class Component internal constructor(
         block: ComponentBuilderBlock? = null
     ): Graph = Graph(application, null, this, null, block)
 
-    /**
-     * @see createGraph
-     */
-    @JvmOverloads
-    @Deprecated(
-        "Use createGraph instead.",
-        ReplaceWith("createGraph(application,block)")
-    )
-    fun init(
-        application: WinterApplication = Winter,
-        block: ComponentBuilderBlock? = null
-    ): Graph = createGraph(application, block)
-
-    internal inline fun forEach(block: (Map.Entry<TypeKey, UnboundService<*, *>>) -> Unit) {
+    internal inline fun forEach(block: (Map.Entry<TypeKey<*, *>, UnboundService<*, *>>) -> Unit) {
         dependencies.forEach(block)
     }
 
-    internal fun keys(): Set<TypeKey> = dependencies.keys
+    internal fun keys(): Set<TypeKey<*, *>> = dependencies.keys
 
-    internal operator fun get(key: TypeKey): UnboundService<*, *>? = dependencies[key]
+    internal operator fun get(key: TypeKey<*, *>): UnboundService<*, *>? = dependencies[key]
 
     internal val size: Int get() = dependencies.size
 
     internal fun isEmpty(): Boolean = dependencies.isEmpty()
 
-    internal fun containsKey(typeKey: TypeKey): Boolean = dependencies.containsKey(typeKey)
+    internal fun containsKey(typeKey: TypeKey<*, *>): Boolean = dependencies.containsKey(typeKey)
 
 }