+++ winter/src/main/kotlin/io/jentz/winter/WinterException.kt
@@ -9,13 +9,13 @@ open class WinterException(message: String?, cause: Throwable? = null) : Excepti
  * Exception that is thrown when a component entry or graph was not found but was requested as
  * non-optional.
  */
-class EntryNotFoundException(val key: TypeKey, message: String) : WinterException(message)
+class EntryNotFoundException(val key: TypeKey<*, *>, message: String) : WinterException(message)
 
 /**
  * Exception that is thrown when an error occurs during dependency resolution.
  */
 class DependencyResolutionException(
-    val key: TypeKey,
+    val key: TypeKey<*, *>,
     message: String,
     cause: Throwable? = null
 ) : WinterException(message, cause)
@@ -24,7 +24,7 @@ class DependencyResolutionException(
  * Exception that is thrown when a cyclic dependency was detected.
  */
 class CyclicDependencyException(
-    val key: TypeKey,
+    val key: TypeKey<*, *>,
     message: String,
     cause: Throwable? = null
 ) : WinterException(message, cause)