+++ winter/src/test/kotlin/io/jentz/winter/assertions.kt
@@ -14,7 +14,7 @@ inline fun <T> expectValueToChange(from: T, to: T, valueProvider: () -> T, block
     if (b != to) fail("Expected change from ${formatValue(from)} to ${formatValue(to)} but was ${formatValue(b)}")
 }
 
-fun Graph.shouldContainService(key: TypeKey, alsoCheckParent: Boolean = false) {
+fun Graph.shouldContainService(key: TypeKey<*, *>, alsoCheckParent: Boolean = false) {
     if (!component.containsKey(key)) {
         val parent = this.parent
         if (parent == null || !alsoCheckParent) {
@@ -24,7 +24,7 @@ fun Graph.shouldContainService(key: TypeKey, alsoCheckParent: Boolean = false) {
     }
 }
 
-fun Graph.shouldNotContainService(key: TypeKey, alsoCheckParent: Boolean = false) {
+fun Graph.shouldNotContainService(key: TypeKey<*, *>, alsoCheckParent: Boolean = false) {
     if (component.containsKey(key)) {
         val parent = this.parent
         if (parent == null || !alsoCheckParent) {
@@ -34,17 +34,17 @@ fun Graph.shouldNotContainService(key: TypeKey, alsoCheckParent: Boolean = false
     }
 }
 
-internal inline fun <reified S : UnboundService<*, *>> Component.shouldContainServiceOfType(key: TypeKey) {
+internal inline fun <reified S : UnboundService<*, *>> Component.shouldContainServiceOfType(key: TypeKey<*, *>) {
     val service = this[key]
             ?: fail("Component was expected to contain service with key <$key> but doesn't")
     if (service !is S) fail("Service with key <$key> was expected to be <${S::class}> but was <${service.javaClass}>.")
 }
 
-internal fun Component.shouldContainService(key: TypeKey) {
+internal fun Component.shouldContainService(key: TypeKey<*, *>) {
     if (!containsKey(key)) fail("Component was expected to contain service with key <$key> but doesn't")
 }
 
-internal fun Component.shouldNotContainService(key: TypeKey) {
+internal fun Component.shouldNotContainService(key: TypeKey<*, *>) {
     if (containsKey(key)) fail("Component wasn't expected to contain service with key <$key> but does.")
 }